stockpulse-0.1.0/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "stockpulse"
version = "0.1.0"
authors = [
  { name = "ghost(Vongchak Jonathan)", email = "vongchaknanbol@gmail.com" },
]
description = "A robust Python package for fetching and managing stock market data with yfinance to avoid frequent rate-limiting."
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
  "yfinance>=0.2.0",
  "pandas>=1.5.0",
  "pyyaml>=6.0",
]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Intended Audience :: Developers",
  "Topic :: Office/Business :: Financial",
]

[project.urls]
Homepage = "https://github.com/jayayjay/stockpulse"
Issues = "https://github.com/jayayjay/stockpulse/issues"
Documentation = "https://stockpulse.readthedocs.io/"